Move courses tree provider and refactor get files in extension
