+++ vscode4teaching-extension/src/components/courses/CoursesTreeProvider.ts
