+++ vscode4teaching-extension/test/integrationSuite/extension.test.ts
@@ -3,13 +3,13 @@ import { afterEach, suite, test } from 'mocha';
 import * as vscode from 'vscode';
 import * as extension from '../../src/extension';
 import * as simple from 'simple-mock';
-import { V4TItem, V4TItemType } from '../../src/components/coursesTreeProvider/V4TItem';
+import { V4TItem, V4TItemType } from '../../src/components/courses/V4TItem';
 import { Course, Exercise, User } from '../../src/model/serverModel/ServerModel';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as rimraf from 'rimraf';
 import * as JSZip from 'JSZip';
-import { UserPick } from '../../src/components/coursesTreeProvider/CoursesTreeProvider';
+import { UserPick } from '../../src/components/courses/CoursesTreeProvider';
 import { APIClient } from "../../src/client/APIClient";
 import { Validators } from '../../src/model/Validators';
 import { CurrentUser } from '../../src/model/CurrentUser';
